\name{area_circle}
\alias{area_circle}
\title{Area of a circle}
\description{Compute the area of the circle by providing the radius.}
\usage{area_circle(radius)}
\arguments{\item{radius}{numeric. The radius must be numeric and greater than zero.}}
\value{numeric}
\examples{area_circle(10)}

% the below are not strictly required (examples neither, but good to include)

\details{Not necessary for this package, but keep this for reference.}
\references{Pythagoras (2000 a.C)}
\author{JV Silva}
\note{R package developed during the workshop.}
\section{Warning}{not much tested yet!}
\seealso{\code{\link{length}}}
\keyword{spatial}
\concept{workshop}
