% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dogs.R
\name{dog_loves}
\alias{dog_loves}
\title{Dog}
\usage{
dog_loves(love = TRUE)
}
\arguments{
\item{love}{Do you love dogs? Defaults to TRUE}
}
\value{

}
\description{
Dog
}
\examples{
dog_loves(Love)
}
\keyword{Dogs}
