% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_columns.R
\name{drop_columnnas}
\alias{drop_columnnas}
\title{Drops NA}
\usage{
drop_columnnas(df, threshold)
}
\arguments{
\item{df}{This is a structured dataframe with any datatypes inside}

\item{threshold}{cutoff for missing data from 0-1}
}
\value{

}
\description{
Drops variables that have too man NA Values
}
\examples{
drop_columnnas( df, 0.1)
}
