% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string specifying the filename that needs to be read.}
}
\value{
A data frame containing a representation of the data in the file.
}
\description{
This is a simple function that reads a file by taking filename as input. If a filename doesn't exist in the working directory then the function
stops with the message "file 'filename' does not exist".
}
\examples{
fars_read("accident_2013.csv.bz2")

}
