% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of data frames.
}
\description{
This function let's us read multiple files by just inputting the years for which we want to read files. It generates a list of data frames,
where each file is a data frame that's part of the list.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}

}
