\name{firstRpackage-package}

\alias{firstRpackage-package}

\alias{firstRpackage}

\docType{package}

\title{\packageTitle{firstRpackage}}

\description{\packageDescription{firstRpackage}}

\details{no description yet}

\author{
\packageAuthor{firstRpackage}
Maintainer: \packageMaintainer{firstRpackage}}
