% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstRPakage.R
\docType{package}
\name{FirstRPackage}
\alias{FirstRPackage}
\alias{FirstRPackage-package}
\title{packr: A package with a minimum example of package creation.}
\description{
This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Megan Magee, Department of Chemistry and Chemical Biology, McMaster University \email{mageem@mcmaster.ca}
}
