% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which you want to create the filename.}
}
\value{
A filename for the given year.
}
\description{
This is a simple function that creates a filename in the ".csv.bz2" format starting with "accident_" for a given year.
}
\examples{
make_filename(2018)

}
