% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean1}
\alias{mean1}
\title{Taking a Mean of a set of values
The purpose of this function is to take the mean of a set of values}
\usage{
mean1(x)
}
\arguments{
\item{x}{the values of a certain parameter}
}
\value{
The mean of a set of values
}
\description{
Taking a Mean of a set of values
The purpose of this function is to take the mean of a set of values
}
\examples{
(data = StandardizationAssayData)
}
