% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mply.R
\name{mply}
\alias{mply}
\title{Multiply two values}
\usage{
mply(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The value of a times b
}
\description{
Multiply two values
}
\examples{
print("Hello, World!")
mply (2,4)
}
