% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstRPakage.R
\docType{data}
\name{StandardizationAssayData}
\alias{StandardizationAssayData}
\title{Standardization Assay Data}
\format{A data frame with 12 rows of data per sample vial and 6 variables
\describe{
  \item{Peak MT}{The Migration times for each peak seen on the UV spectra}
  \item{Peak Area}{Peak areas for each peak seen on the UV spectra}
  \item{Relative PA}{Relative Peak Areas, which are peak areas which have been normalized with respect to the Li Peak area}
  \item{Identity}{The identiry/ ion which is displaying a peak for each UV spectra signal}
  \item{Data File}{Id for the data file with the experiments}
  \item{Samplle Vial}{The sample vial number}
}}
\source{
Megan Magee: CE-UV instrumentation results processed data Created: September 25, 2019
}
\usage{
data(StandardizationAssayData)
}
\description{
A dataset containing information about common urinary metabolites and their annalysis under CE-UV
which produced information pertaining to the ions peak migration times, peak areas and relative peak areas with respect to Lithium internal standard ion
}
\examples{
data(StandardizationAssayData)
}
\keyword{datasets}
