% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindowBacktesting.R
\name{window_backtesting}
\alias{window_backtesting}
\title{Window Backtesting Scheme}
\usage{
window_backtesting(
  model,
  data,
  orig,
  h,
  xreg = NULL,
  fixed = NULL,
  inc.mean = TRUE,
  reest = 1
)
}
\arguments{
\item{model}{this is your model}

\item{data}{data type}

\item{reest}{}
}
\value{

}
\description{
Window Backtesting Scheme
}
